+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="0.0.2"></a>
+## 0.0.2 (2025-10-17)
+
+### Features
+
+Added an application that exposes a REST api to convert an interger into its written form
+
+> Note: to keep the two running apps independant, the source code of NumbersToWords class has been duplicated  :/
+
+
 <a name="0.0.1"></a>
 ## 0.0.1 (2025-10-16)